Incorporated connection pooling and modified Controllers to mainly rely on built-in ReactiveRepository calls for sake of performance

+++ src/main/kotlin/hvalfangst/cars/controller/CarDetailsController.kt
@@ -13,7 +13,7 @@ import reactor.core.publisher.Mono
 class CarDetailsController(private val carDetailsService: CarDetailsService) {
 
     @GetMapping("/{carId}")
-    fun getCarDetails(@PathVariable carId: Int): Mono<CarDetails> {
+    fun getCarDetails(@PathVariable carId: Long): Mono<CarDetails> {
         return carDetailsService.getCarDetails(carId)
     }
 }
